Image width and height now in input text file and scaling changed
Since width/height are now in input, we scale the image up
only as much as necessary so both the width and height in the
window are filled.

Also, hat.py program refactored. `match_rev` is eliminated and only
the chiralities are compared to determine which direction the new
figure should be drawn. Furthermore, chirality is determined from
the color in the config file. More partially visible tiles are added
around the edge.

For honeycomb.py, `SCALING` is now set to 1, as there is no need
to scale in this program anymore.