Set crop values using new `set_crop_values` function
The values can now be set after initialization. This is useful
because we will sometimes want to set the values to values that
might not equal a tile point.

For the same reason, values are now set by coordinate and not
by passing the `PointRef` namedtuple. In fact, with this change,
there is no need for the `PointRef` namedtuple, and it is removed.
`AllTiles.get_pt` now takes `tile_id` and `edge` parameters instead
of `PointRef`.