Add tilings for six other hat-like variants
The hat_config.txt is also changed to use 'BASED_ON_TILING' in the
footnote. This string is then replaced with a description of the
tile type when making the footnotes in `hatlike.py`.

The degenerate hat-like tiles, i.e., those with one edge of length
zero, are still not supported.