For spectre tilings, indicate in coordinate file whether to curve edges
Also, now create two spectre tilings, one with this indicator and one
without. Update `show_progress.py` to use this new indicator.
Add `pass_through` attribute to `AllTiles` to store this parameter
and similar parameters that don't need to be acted on by the program
that makes the coordinate files (except for passing the value through
to the output file).